perf: refactor interactive submodule
